add file
In a CTE query, combine the data from the above tables to get: ad_date, url_parameters - part of the URL from the campaign link that includes UTM parameters, spend, impressions, reach, clicks, leads, value - metrics of campaigns and ad sets on the respective days. If the metric value is missing in the table (i.e. null), set the value to zero (i.e. 0). From the resulting CTE, make a sample: ad_date, utm_campaign - the value of the utm_campaign parameter from the utm_parameters field that meets the following conditions: it is lowercase, if the value of utm_campaign in utm_parameters is 'nan', then it must be empty (i.e. null) in the resulting table (use the substring function with a regular expression). Total cost, number of impressions, number of clicks, as well as total Conversion Value on the corresponding date for the corresponding campaign; CTR, CPC, CPM, ROMI on the corresponding date for the corresponding campaign. At the same time, do not use WHERE, but avoid the division by zero error by using the CASE operator.
Additional task: Encoding strings in the URL. In the values of the utm_parameters field, Cyrillic letters are encoded into a url string. For example, we have utm_parameters = 'utm_source=facebook&utm_mediun=ppc&utm_campaign=trend'.  In the query you created for the main task of this assignment, decode the value of utm_campaign by creating a temporary function. Find the function code online.