add file
Use the CTE from the previous task in a new (second) CTE to create a sample with the following data: ad_month - the first day of the month of the ad's display date (obtained from ad_date); utm_campaign, total cost, number of impressions, number of clicks, conversion value, CTR, CPC, CPM, ROMI - the same fields with the same conditions as in the previous task. Make the resulting selection with the following fields: ad_month; utm_campaign, total cost, number of impressions, number of clicks, conversion value, CTR, CPC, CPM, ROMI. For each utm_campaign in each month, add a new field: 'difference in CPM, CTR, and ROMI' in the current month compared to the previous month in percentage.